finished create song
