+++ app/seeds/__init__.py
@@ -1,8 +1,8 @@
 from flask.cli import AppGroup
-from .users import seed_users, undo_users
-from .songs import seed_songs, undo_songs
-from .comments import seed_comments, undo_comments
-from .liked_songs import seed_liked_songs, undo_liked_songs
+from .user_seeds import seed_users, undo_users
+from .song_seeds import seed_songs, undo_songs
+from .comments_seeds import seed_comments, undo_comments
+from .liked_songs_seeds import seed_liked_songs, undo_liked_songs
 
 from app.models.db import db, environment, SCHEMA
 
@@ -24,7 +24,10 @@ def seed():
         db.session.execute(f"TRUNCATE table {SCHEMA}.comments RESTART IDENTITY CASCADE;")
         db.session.execute(f"TRUNCATE table {SCHEMA}.liked_songs RESTART IDENTITY CASCADE;")
         db.session.commit()
-
+        undo_liked_songs()
+        undo_comments()
+        undo_songs()
+        undo_users()
 
     seed_users()
     seed_songs()
