+++ react-vite/src/components/CreateSongForm/CreateSongForm.css
@@ -1,0 +1,12 @@
+.error-cont{
+    max-height: 40px;
+}
+
+/* FORM CONTAINER*/
+/* .created-song-form-cont{
+    display: flex;
+    flex-direction: column;
+    gap:50px;
+    height:100%;
+} */
+/******/
